+++ packages/amplify-category-predictions/provider-utils/awscloudformation/prediction-category-walkthroughs/infer-walkthrough.js
@@ -2,7 +2,7 @@ import open from 'open';
 import inferAssets from '../assets/inferQuestions';
 import getAllDefaults from '../default-values/infer-defaults';
 import regionMapper from '../assets/regionMapping';
-
+import { ResourceAlreadyExistsError, ResourceDoesNotExistError, exitOnNextTick } from 'amplify-cli-core';
 const inquirer = require('inquirer');
 const path = require('path');
 const fs = require('fs-extra');
@@ -26,11 +26,14 @@ async function addWalkthrough(context) {
         await add(context);
       } catch (e) {
         context.print.error('The Auth plugin is not installed in the CLI. You need to install it to use this feature');
+        context.usageData.emitError(e);
+        exitOnNextTick(1);
         break;
       }
       break;
     } else {
-      process.exit(0);
+      context.usageData.emitSuccess();
+      exitOnNextTick(0);
     }
   }
 
@@ -52,8 +55,10 @@ async function updateWalkthrough(context) {
     }
   });
   if (predictionsResources.length === 0) {
-    context.print.error('No resources to update. You need to add a resource.');
-    process.exit(0);
+    const errMessage = 'No resources to update. You need to add a resource.';
+    context.print.error(errMessage);
+    context.usageData.emitError(new ResourceDoesNotExistError(errMessage));
+    exitOnNextTick(0);
     return;
   }
   let resourceObj = predictionsResources[0].value;
@@ -96,8 +101,10 @@ async function configure(context, resourceObj) {
     // check if that type is already created
     const resourceType = resourceAlreadyExists(context, answers.inferType);
     if (resourceType) {
-      context.print.warning(`${resourceType} has already been added to this project.`);
-      process.exit(0);
+      const errMessage = `${resourceType} has already been added to this project.`;
+      context.print.warning(errMessage);
+      context.usageData.emitError(new ResourceAlreadyExistsError(errMessage));
+      exitOnNextTick(0);
     }
 
     Object.assign(answers, await inquirer.prompt(inferAssets.setup.name(`${answers.inferType}${defaultValues.resourceName}`)));
@@ -244,8 +251,10 @@ async function getEndpoints(context, questionObj, params) {
     endpointMap[endpoint.EndpointName] = { endpointName: endpoint.EndpointName, endpointARN: endpoint.EndpointArn };
   });
   if (endpoints.length < 1) {
-    context.print.error('No existing endpoints!');
-    process.exit(0);
+    const errMessage = 'No existing endpoints!';
+    context.print.error(errMessage);
+    context.usageData.emitError(new ResourceDoesNotExistError(errMessage));
+    exitOnNextTick(0);
   }
   const { endpoint } = await inquirer.prompt(questionObj.importPrompt({ ...params, endpoints }));
   return endpointMap[endpoint];