+++ packages/amplify-category-storage/provider-utils/awscloudformation/service-walkthroughs/s3-walkthrough.js
@@ -4,6 +4,7 @@ const fs = require('fs-extra');
 const _ = require('lodash');
 const uuid = require('uuid');
 const { ServiceName: FunctionServiceName } = require('amplify-category-function');
+const { ResourceDoesNotExistError, ResourceAlreadyExistsError, exitOnNextTick } = require('amplify-cli-core');
 
 const category = 'storage';
 const parametersFileName = 'parameters.json';
@@ -31,18 +32,22 @@ async function addWalkthrough(context, defaultValuesFilename, serviceMetadata, o
         await add(context);
       } catch (e) {
         context.print.error('The Auth plugin is not installed in the CLI. You need to install it to use this feature');
-        break;
+        context.usageData.emitError(e);
+        exitOnNextTick(1);
       }
       break;
     } else {
-      process.exit(0);
+      context.usageData.emitSuccess();
+      exitOnNextTick(0);
     }
   }
   const resourceName = resourceAlreadyExists(context);
 
   if (resourceName) {
-    context.print.warning('Amazon S3 storage was already added to your project.');
-    process.exit(0);
+    const errMessage = 'Amazon S3 storage was already added to your project.';
+    context.print.warning(errMessage);
+    context.usageData.emitError(new ResourceAlreadyExistsError(errMessage));
+    exitOnNextTick(0);
   } else {
     return await configure(context, defaultValuesFilename, serviceMetadata, undefined, options);
   }
@@ -55,14 +60,16 @@ function updateWalkthrough(context, defaultValuesFilename, serviceMetada) {
   const storageResources = {};
 
   Object.keys(amplifyMeta[category]).forEach(resourceName => {
-    if (amplifyMeta[category][resourceName].service === serviceName) {
+    if (amplifyMeta[category][resourceName].service === serviceName && !!amplifyMeta[category][resourceName].providerPlugin) {
       storageResources[resourceName] = amplifyMeta[category][resourceName];
     }
   });
 
   if (Object.keys(storageResources).length === 0) {
-    context.print.error('No resources to update. You need to add a resource.');
-    process.exit(0);
+    const errMessage = 'No resources to update. You need to add a resource.';
+    context.print.error(errMessage);
+    context.usageData.emitError(new ResourceDoesNotExistError(errMessage));
+    exitOnNextTick(0);
     return;
   }
   const [resourceName] = Object.keys(storageResources);