+++ packages/amplify-e2e-tests/src/__tests__/auth_1.test.ts
@@ -1,33 +1,17 @@
-import * as fs from 'fs-extra';
-import {
-  initJSProjectWithProfile,
-  initAndroidProjectWithProfile,
-  initIosProjectWithProfile,
-  deleteProject,
-  amplifyPushAuth,
-  amplifyPush,
-} from 'amplify-e2e-core';
+import { initJSProjectWithProfile, deleteProject, amplifyPushAuth, amplifyPush } from 'amplify-e2e-core';
 import {
   addAuthWithDefault,
+  runAmplifyAuthConsole,
   removeAuthWithDefault,
   addAuthWithDefaultSocial,
   addAuthWithGroupTrigger,
   addAuthWithRecaptchaTrigger,
-  addAuthWithCustomTrigger,
-  addAuthWithSignInSignOutUrl,
-  updateAuthWithoutCustomTrigger,
-  updateAuthRemoveRecaptchaTrigger,
-  updateAuthSignInSignOutUrl,
   addAuthViaAPIWithTrigger,
-  addAuthWithMaxOptions,
-  addAuthUserPoolOnly,
 } from 'amplify-e2e-core';
 import {
   createNewProjectDir,
   deleteProjectDir,
   getProjectMeta,
-  getAwsAndroidConfig,
-  getAwsIOSConfig,
   getUserPool,
   getUserPoolClients,
   getLambdaFunction,
@@ -52,6 +36,7 @@ describe('amplify add auth...', () => {
     await initJSProjectWithProfile(projRoot, defaultsSettings);
     await addAuthWithDefault(projRoot, {});
     await amplifyPushAuth(projRoot);
+    await runAmplifyAuthConsole(projRoot);
     const meta = getProjectMeta(projRoot);
     const id = Object.keys(meta.auth).map(key => meta.auth[key])[0].output.UserPoolId;
     const userPool = await getUserPool(id, meta.providers.awscloudformation.Region);